harden(rate-limit): source time from the store, not the app clock
Removes the last flagged shortcut. Refill timing now comes from ONE
authoritative clock per store instead of each app instance's Date.now(),
so app-server clock skew can no longer affect the token math — making
'distributed-correct' literally true rather than 'true assuming NTP'.

- Postgres: clock_timestamp() evaluated once in the VALUES clause and
  reused via EXCLUDED.last_refill_ms in the UPDATE (a second call would
  return a later value and desync the two places now is used). Cleanup
  DELETE uses DB time too. last_refill_ms column widened to DOUBLE
  PRECISION to keep sub-ms resolution.
- Redis: redis.call('TIME') at the top of the Lua, folded seconds+usec to
  ms (effects-replication, default since Redis 5, permits the pre-write
  read; we target 7).
- MemoryStore: unchanged — one process, one clock, no skew possible.

No dependency change: both are edits to SQL/Lua strings already flowing
through the existing store interfaces (deps still {}). Unit-test emulators
updated to model the single server clock. rate-limit 17 tests + 2 real-DB
integration (skipped without a database); public API unchanged.

+++ packages/rate-limit/src/migrations/sql/001_rate_limits.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS fonderie_rate_limits (
 	key            TEXT             NOT NULL PRIMARY KEY,
 	tokens         DOUBLE PRECISION NOT NULL,
-	last_refill_ms BIGINT           NOT NULL,
+	last_refill_ms DOUBLE PRECISION NOT NULL,
 	granted        BOOLEAN          NOT NULL DEFAULT TRUE
 );
